Rewrite AsyncTelnet for performance
We noticed that servers dumping loads of data and no commands into VSPC
would be able to fill up their socket's buffer easily, because VSPC was
not able to handle data fast enough. This came from the parsing and all
the string concatenation stemming from handling a single char at a time.

We rewrote the whole AsyncTelnet handling to make use of string indizes
and keep the creation of new strings (i.e. string concatenation) at a
minimum. This also shows in the performance tests running on my laptop
one after another:

	vspc.tests.test_telnet.AsyncTelnetTest.test_performance_mixed                3.979
	vspc.tests.test_telnet.AsyncTelnetTest.test_performance_lots_of_commands     1.424
	vspc.tests.test_telnet.AsyncTelnetTest.test_performance_lots_of_escaped_iac  0.051
	vspc.tests.test_telnet.AsyncTelnetTest.test_performance_lots_of_data         0.007