Revert "Added aiofiles to read/write files"
This reverts commit bbff0b26e912fa41b710d3efeffc5fb636534fa5.

Pushing every little tiny update into an own thread with
asyncio.run_in_executor() is just too slow to get anything down to disk.
We have to find a better solution for that and revert the change for
now.

The performance tests for writing improve to < 1s (single) and < 20s
(multiple) from ~8s (single) and ~105s (multiple).