tests: Add sleep to simulate socket read
When mocking AsyncTelnet.read_some(), we now use a coroutine and add an
async.sleep() into the mix to let other coroutines run. This simulates
reading from a socket and shows more similar behavior to the real world,
with sockets having data come in parallel and thus writes happening in
parallel.