Handle file writes in background thread
When starting the VspcServer, we also start another thread:
BackgroundWriter. This thread is responsible for synchronously writing
the logs to file. We decouple it from the main thread to not let hanging
file writes stop the handling for telnet messages, as this can prohibit
vMotions (the ESXi needs the ACK for a vMotion in <= 5s to actually
start the vMotion).

Since we rely on another thread to be working properly, we introduce a
check method into the main thread. Finding the BackgroundWriter not
alive, we will exit the whole application and hope for k8s or other
means to restart it.

This also increases performance tests speed to < 12s (multiple) and
keeps the single test roughly the same.

We also add a simple writing test here for simple debugging.