Estimate available data instead of counting
Since we run in two threads, the `deque`s holding the to-be-written data
can change while we iterate over them to count the amount of data
available for the decision if we want to start flushing the data.

Therefore, we explicitly count the sum of data we estimate to be in the
`deque` objects. We say "estimate" here, because the whole counting we
do is not thread-safe. Since we flush the queues every 5s anyways, we do
not need too much accuracy.