Do not log every _flush_queues_to_disk()
The amount of logging done by the regular (every 5s) call to
`_flush_queues_to_disk()` isn't really helpful and clutters the view on
what's actually happening in the vspc service. Therefore, we remove it.

We do not change it to show the amount of bytes written, when there was
actual writing, as big regions with large amounts of VMs will probably
write something all the time, so this information will not be too
helpful, I'd guess.